refactor(core): do not leak vm specific implementations from `vm` crate.
