+++ crates/blockchain/payload.rs
@@ -19,8 +19,7 @@ use ethrex_common::{
 };
 
 use ethrex_vm::{
-    backends::{Evm, EvmEngine},
-    EvmError,
+    EvmError, {Evm, EvmEngine},
 };
 
 use ethrex_rlp::encode::RLPEncode;
