+++ crates/networking/rpc/eth/transaction.rs
@@ -22,7 +22,7 @@ use ethrex_common::{
 use ethrex_rlp::encode::RLPEncode;
 use ethrex_storage::Store;
 
-use ethrex_vm::{backends::Evm, execution_result::ExecutionResult};
+use ethrex_vm::{Evm, ExecutionResult};
 use serde::Serialize;
 
 use serde_json::Value;
