+++ crates/vm/Cargo.toml
@@ -41,6 +41,7 @@ path = "./lib.rs"
 
 [features]
 default = []
+internal = []
 l2 = ["ethrex-levm/l2"]
 c-kzg = ["revm/c-kzg", "ethrex-levm/c-kzg", "ethrex-common/c-kzg"]
 blst = ["revm/blst"]
