+++ crates/vm/backends/revm/mod.rs
@@ -9,9 +9,9 @@ use crate::constants::{
     BEACON_ROOTS_ADDRESS, CONSOLIDATION_REQUEST_PREDEPLOY_ADDRESS, HISTORY_STORAGE_ADDRESS,
     SYSTEM_ADDRESS, WITHDRAWAL_REQUEST_PREDEPLOY_ADDRESS,
 };
+use crate::errors::EvmError;
 use crate::execution_result::ExecutionResult;
-use crate::spec::spec_id;
-use crate::EvmError;
+use crate::helpers::spec_id;
 use db::EvmState;
 use ethrex_common::types::AccountInfo;
 use ethrex_common::{BigEndianHash, H256, U256};
