+++ crates/vm/helpers.rs
@@ -1,8 +1,5 @@
-// use crate::backends::revm::*;
-
-// Export needed types
 use ethrex_common::types::{ChainConfig, Fork};
-pub use revm::primitives::{Address as RevmAddress, SpecId, U256 as RevmU256};
+pub use revm::primitives::SpecId;
 
 /// Returns the spec id according to the block timestamp and the stored chain config
 /// WARNING: Assumes at least Merge fork is active
@@ -33,4 +30,15 @@ pub fn fork_to_spec_id(fork: Fork) -> SpecId {
         Fork::Prague => SpecId::PRAGUE,
         Fork::Osaka => SpecId::OSAKA,
     }
+}
+
+use ethrex_common::Address;
+
+pub fn create_contract_address(from: Address, nonce: u64) -> Address {
+    Address::from_slice(
+        revm::primitives::Address(from.0.into())
+            .create(nonce)
+            .0
+            .as_ref(),
+    )
 }
